refactor(youtube/integrations): rewrite codes

+++ app/src/main/java/app/revanced/integrations/swipecontrols/views/SwipeControlsOverlayLayout.kt
@@ -77,7 +77,7 @@ class SwipeControlsOverlayLayout(
                 setColor(config.overlayTextBackgroundColor)
             }
             setTextColor(config.overlayForegroundColor)
-            setTextSize(TypedValue.COMPLEX_UNIT_SP, config.overlayTextSize)
+            setTextSize(TypedValue.COMPLEX_UNIT_SP, config.overlayTextSize.toFloat())
             compoundDrawablePadding = compoundIconPadding
             visibility = GONE
         }
